requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
pydantic
pydantic-settings
bcrypt
python-jose[cryptography]
email-validator
sqlalchemy[asyncio]
asyncpg
